+++ packages/core/src/services/aws-saml-assertion-extraction-service.ts
@@ -16,7 +16,8 @@ const authenticationUrlRegexes = new Map([
       /^https:\/\/login\.microsoftonline\.com\/*.*\/oauth2\/authorize.*/,
       /^https:\/\/.+\.auth0\.com\/samlp\/.+/,
       /^https:\/\/.*[/auth]?\/realms\/.*\/protocol\/saml\/clients\/.*/,
-      /^https:\/\/sso\.jumpcloud\.com\/saml2\/aws/,
+      /^https:\/\/console\.jumpcloud\.com\/login.*/,
+      /^https:\/\/accounts\.google\.com\/AccountChooser.*/, 
     ],
   ],
 ]);